Move files into individual packages
