+++ src/main/java/duke/action/Action.java
@@ -1,9 +1,13 @@
+package duke.action;
+
+import duke.Response;
+
 /**
  * The Action interface represents an action to be performed in the Duke application.
  */
 public interface Action {
     /**
-     * Executes the specified action and returns the resulting Response.
+     * Executes the specified action and returns the resulting duke.Response.
      */
     Response execute();
 }
