+++ src/main/java/duke/action/AddTask.java
@@ -1,14 +1,20 @@
+package duke.action;
+
+import duke.Response;
+import duke.task.Task;
+import duke.task.TaskCollection;
+
 /**
- * AddTask is an Action that adds a Task to a TaskCollection.
+ * duke.action.AddTask is an Action that adds a duke.task.Task to a duke.task.TaskCollection.
  */
 public class AddTask implements Action {
     private final Task task;
     private final TaskCollection taskCollection;
 
     /**
-     * Creates an Action AddTask that adds a Task into a TaskCollection.
-     * @param task The Task to be added.
-     * @param taskCollection The TaskCollection to add the Task into.
+     * Creates an Action duke.action.AddTask that adds a duke.task.Task into a duke.task.TaskCollection.
+     * @param task The duke.task.Task to be added.
+     * @param taskCollection The duke.task.TaskCollection to add the duke.task.Task into.
      */
     public AddTask(Task task, TaskCollection taskCollection) {
         this.task = task;
@@ -16,8 +22,8 @@ public class AddTask implements Action {
     }
 
     /**
-     * Adds the Task into the TaskCollection.
-     * @return The Response.
+     * Adds the duke.task.Task into the duke.task.TaskCollection.
+     * @return The duke.Response.
      */
     @Override
     public Response execute() {
@@ -26,8 +32,8 @@ public class AddTask implements Action {
     }
 
     /**
-     * Returns the Response for successfully adding the Task to the TaskCollection.
-     * @return The Response.
+     * Returns the duke.Response for successfully adding the duke.task.Task to the duke.task.TaskCollection.
+     * @return The duke.Response.
      */
     private Response respond() {
         return new Response(new String[]{
