+++ src/main/java/duke/action/CompleteTask.java
@@ -1,30 +1,35 @@
+package duke.action;
+
+import duke.Response;
+import duke.task.Task;
+
 /**
- * CompleteTask is an Action that marks a Task as done.
+ * duke.action.CompleteTask is an Action that marks a duke.task.Task as done.
  */
 public class CompleteTask implements Action {
     private final Task task;
 
     /**
-     * Creates an Action CompleteTask that marks a Task as done.
-     * @param task The Task to be marked as done.
+     * Creates an Action duke.action.CompleteTask that marks a duke.task.Task as done.
+     * @param task The duke.task.Task to be marked as done.
      */
     public CompleteTask(Task task) {
         this.task = task;
     }
 
     /**
-     * Marks the Task as done.
-     * @return The Response.
+     * Marks the duke.task.Task as done.
+     * @return The duke.Response.
      */
     @Override
     public Response execute() {
-        this.task.isDone = true;
+        this.task.markDone();
         return this.respond();
     }
 
     /**
-     * Returns the Response for marking the Task as done.
-     * @return The Response.
+     * Returns the duke.Response for marking the duke.task.Task as done.
+     * @return The duke.Response.
      */
     private Response respond() {
         return new Response(new String[]{
