+++ src/main/java/duke/request/Command.java
@@ -1,5 +1,9 @@
+package duke.request;
+
+import duke.exception.UserException;
+
 /**
- * A Command represents a string in a Request that instructs the application to perform an Action.
+ * A duke.request.Command represents a string in a duke.request.Request that instructs the application to perform an Action.
  */
 public enum Command {
     BYE("bye"),
@@ -11,22 +15,22 @@ public enum Command {
     TODO("todo");
 
     /**
-     * The name of the Command.
+     * The name of the duke.request.Command.
      */
     private final String name;
 
     /**
-     * Creates a Command with the specified name.
-     * @param name Name of the Command.
+     * Creates a duke.request.Command with the specified name.
+     * @param name Name of the duke.request.Command.
      */
     Command(String name) {
         this.name = name;
     }
 
     /**
-     * Parses the input String to determine the related Command.
+     * Parses the input String to determine the related duke.request.Command.
      * @param commandString The input command String.
-     * @return The Command related to the command String.
+     * @return The duke.request.Command related to the command String.
      * @throws UserException If the command String is invalid.
      */
     public static Command parseFrom(String commandString) throws UserException {
