+++ src/main/java/duke/request/DoneRequest.java
@@ -1,14 +1,22 @@
+package duke.request;
+
+import duke.task.Task;
+import duke.task.TaskCollection;
+import duke.exception.UserException;
+import duke.action.Action;
+import duke.action.CompleteTask;
+
 import java.lang.NumberFormatException;
 
 /**
- * DoneRequest represents a request from the user to mark a Task as done in the application.
+ * duke.request.DoneRequest represents a request from the user to mark a duke.task.Task as done in the application.
  */
 public class DoneRequest extends TaskCollectionRequest {
     private final int taskId;
 
     /**
-     * Creates a DoneRequest.
-     * @param taskCollection The target TaskCollection.
+     * Creates a duke.request.DoneRequest.
+     * @param taskCollection The target duke.task.TaskCollection.
      * @param requestString The request String.
      * @throws UserException If the request String is invalid.
      */
@@ -22,7 +30,7 @@ public class DoneRequest extends TaskCollectionRequest {
     }
 
     /**
-     * Gets the Action the DoneRequest requests to execute.
+     * Gets the Action the duke.request.DoneRequest requests to execute.
      * @return The Action to be executed.
      */
     @Override
