+++ src/main/java/duke/task/Deadline.java
@@ -1,16 +1,18 @@
+package duke.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
 /**
- * Deadline is a task that needs to be done before a specific date/time.
+ * duke.task.Deadline is a task that needs to be done before a specific date/time.
  */
 public class Deadline extends Task {
     protected LocalDate by;
 
     /**
-     * Creates a Deadline with the provided description and date/time.
-     * @param description The description of the Deadline task.
-     * @param by The date/time the Deadline needs to be done before.
+     * Creates a duke.task.Deadline with the provided description and date/time.
+     * @param description The description of the duke.task.Deadline task.
+     * @param by The date/time the duke.task.Deadline needs to be done before.
      */
     public Deadline(String description, String by) {
         super(description);
@@ -18,8 +20,8 @@ public class Deadline extends Task {
     }
 
     /**
-     * Converts the Deadline into a String that represents the Deadline.
-     * @return The String representation of a Deadline.
+     * Converts the duke.task.Deadline into a String that represents the duke.task.Deadline.
+     * @return The String representation of a duke.task.Deadline.
      */
     @Override
     public String toString() {
@@ -27,7 +29,7 @@ public class Deadline extends Task {
     }
 
     /**
-     * Converts the Deadline into a String to be stored in Storage.
+     * Converts the duke.task.Deadline into a String to be stored in duke.Storage.
      * @return String to be stored
      */
     @Override
