+++ src/main/java/duke/task/ToDo.java
@@ -1,18 +1,23 @@
+package duke.task;
+
+import duke.task.Task;
+import duke.task.TaskType;
+
 /**
  * Todo is a task without any date/time attached to it.
  */
 public class ToDo extends Task {
     /**
-     * Creates a ToDo with the provided description.
-     * @param description The description of the ToDo.
+     * Creates a duke.task.ToDo with the provided description.
+     * @param description The description of the duke.task.ToDo.
      */
     public ToDo(String description) {
         super(description);
     }
 
     /**
-     * Converts the ToDo into a String that represents the ToDo.
-     * @return The String representation of a ToDo.
+     * Converts the duke.task.ToDo into a String that represents the duke.task.ToDo.
+     * @return The String representation of a duke.task.ToDo.
      */
     @Override
     public String toString() {
@@ -20,7 +25,7 @@ public class ToDo extends Task {
     }
 
     /**
-     * Converts the Event into a String to be stored in Storage.
+     * Converts the duke.task.Event into a String to be stored in duke.Storage.
      * @return String to be stored
      */
     @Override
